+++ src/components/FolderWatcher.vue
@@ -3,12 +3,14 @@ import { open } from '@tauri-apps/plugin-dialog';
 import { watch, WatchEvent } from '@tauri-apps/plugin-fs';
 import { useLogStore } from '../stores/logs';
 import { useStatusStore } from '../stores/status';
-import { fileList, isMultiGamePgn, isWrite, lichessApiClient, openPath, uploadFolderToRound } from '../utils';
+import { fileList, isMultiGamePgn, uploadFolderToRound } from '../upload';
+import { isWrite, openPath } from '../utils';
 import { computed } from 'vue';
 import debounce from 'debounce';
 import { sep } from '@tauri-apps/api/path';
 import { ref } from 'vue';
 import { BroadcastRound } from '../types';
+import { lichessApiClient } from '../client';
 
 const logs = useLogStore();
 const status = useStatusStore();
